pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent"
version = "0.1.0"
description = "Backend service for Playbook agent runtime"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic>=2.10",
    "pydantic-settings>=2.6",
    "litellm>=1.50",
    "cryptography>=44.0",
    "python-multipart>=0.0.18",
    "bcrypt>=4.0",
    "python-jose[cryptography]>=3.3",
    "email-validator>=2.3.0",
    "httpx>=0.28",
    "crawl4ai>=0.7.6",
    "markitdown>=0.1.3",
    "pypdf>=5.2",
    "python-docx>=1.1",
    "python-pptx>=1.0",
    "striprtf>=0.0.29",
    "chardet>=5.2,<6",
    "packaging>=24.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]
